fix: PostgreSQL 동적 파라미터 정리 + 버그 수정
[Phase 1] 동적 파라미터 빌드 수정 (8개 파일):
- $${params.length + 1} → ? 변환
- $${queryParams.length} → ? 변환
- ANY(string_to_array) → INSTR 변환
- INTERVAL '1 day' 제거
- ::int, ::date 타입 캐스팅 제거
- NULLS LAST 제거
- COUNT(*) FILTER → SUM(CASE) 변환

[Phase 2] 추가 버그 수정 (3개):
- airlines/test-callsigns: 파라미터 불일치 (1→2개)
- actions/[id]: UPDATE rows.length 오류 수정
- actions/[id]: 응답 status 하드코딩 제거

Match Rate: 87% → 95% (예상)

+++ src/app/api/announcements/[id]/route.ts
@@ -71,7 +71,7 @@ export async function GET(
           WHEN start_date <= CURRENT_TIMESTAMP AND end_date >= CURRENT_TIMESTAMP THEN 'active'
           ELSE 'expired'
         END as status,
-        (SELECT COUNT(*)::int FROM announcement_views WHERE announcement_id = ?)::int as "viewCount"
+        (SELECT COUNT(*) FROM announcement_views WHERE announcement_id = ?) as "viewCount"
       FROM announcements
       WHERE id = ?
         AND is_active = true
